move analytics admin bar test to new module spec
