+++ tests/e2e/specs/modules/search-console/fixtures/admin-bar.js
@@ -506,388 +506,3 @@ export const searchConsole = {
 		},
 	],
 };
-export const analytics = {
-	'modules::analytics::report::db20ba9afa3000cd79e2888048a1700c': [
-		{
-			nextPageToken: '10',
-			columnHeader: {
-				dimensions: [
-					'ga:date',
-				],
-				metricHeader: {
-					metricHeaderEntries: [
-						{
-							name: 'Users',
-							type: 'INTEGER',
-						},
-						{
-							name: 'Sessions',
-							type: 'INTEGER',
-						},
-						{
-							name: 'Bounce Rate',
-							type: 'PERCENT',
-						},
-						{
-							name: 'Average Session Duration',
-							type: 'TIME',
-						},
-						{
-							name: 'Goal Completions',
-							type: 'INTEGER',
-						},
-						{
-							name: 'Pageviews',
-							type: 'INTEGER',
-						},
-					],
-				},
-			},
-			data: {
-				dataLastRefreshed: null,
-				isDataGolden: null,
-				rowCount: 54,
-				samplesReadCounts: null,
-				samplingSpaceSizes: null,
-				rows: [
-					{
-						dimensions: [
-							'20190729',
-						],
-						metrics: [
-							{
-								values: [
-									'0',
-									'0',
-									'0.0',
-									'0.0',
-									'0',
-									'0',
-								],
-							},
-							{
-								values: [
-									'9',
-									'0',
-									'0.0',
-									'0.0',
-									'0',
-									'11',
-								],
-							},
-						],
-					},
-					{
-						dimensions: [
-							'20190730',
-						],
-						metrics: [
-							{
-								values: [
-									'0',
-									'0',
-									'0.0',
-									'0.0',
-									'0',
-									'0',
-								],
-							},
-							{
-								values: [
-									'7',
-									'3',
-									'66.66666666666666',
-									'4.0',
-									'0',
-									'20',
-								],
-							},
-						],
-					},
-					{
-						dimensions: [
-							'20190731',
-						],
-						metrics: [
-							{
-								values: [
-									'0',
-									'0',
-									'0.0',
-									'0.0',
-									'0',
-									'0',
-								],
-							},
-							{
-								values: [
-									'8',
-									'3',
-									'66.66666666666666',
-									'20.0',
-									'0',
-									'9',
-								],
-							},
-						],
-					},
-					{
-						dimensions: [
-							'20190801',
-						],
-						metrics: [
-							{
-								values: [
-									'0',
-									'0',
-									'0.0',
-									'0.0',
-									'0',
-									'0',
-								],
-							},
-							{
-								values: [
-									'3',
-									'1',
-									'100.0',
-									'0.0',
-									'0',
-									'5',
-								],
-							},
-						],
-					},
-					{
-						dimensions: [
-							'20190802',
-						],
-						metrics: [
-							{
-								values: [
-									'0',
-									'0',
-									'0.0',
-									'0.0',
-									'0',
-									'0',
-								],
-							},
-							{
-								values: [
-									'3',
-									'1',
-									'0.0',
-									'5.0',
-									'0',
-									'4',
-								],
-							},
-						],
-					},
-					{
-						dimensions: [
-							'20190803',
-						],
-						metrics: [
-							{
-								values: [
-									'0',
-									'0',
-									'0.0',
-									'0.0',
-									'0',
-									'0',
-								],
-							},
-							{
-								values: [
-									'3',
-									'1',
-									'100.0',
-									'0.0',
-									'0',
-									'5',
-								],
-							},
-						],
-					},
-					{
-						dimensions: [
-							'20190804',
-						],
-						metrics: [
-							{
-								values: [
-									'0',
-									'0',
-									'0.0',
-									'0.0',
-									'0',
-									'0',
-								],
-							},
-							{
-								values: [
-									'1',
-									'0',
-									'0.0',
-									'0.0',
-									'0',
-									'1',
-								],
-							},
-						],
-					},
-					{
-						dimensions: [
-							'20190805',
-						],
-						metrics: [
-							{
-								values: [
-									'0',
-									'0',
-									'0.0',
-									'0.0',
-									'0',
-									'0',
-								],
-							},
-							{
-								values: [
-									'4',
-									'0',
-									'0.0',
-									'0.0',
-									'0',
-									'5',
-								],
-							},
-						],
-					},
-					{
-						dimensions: [
-							'20190806',
-						],
-						metrics: [
-							{
-								values: [
-									'0',
-									'0',
-									'0.0',
-									'0.0',
-									'0',
-									'0',
-								],
-							},
-							{
-								values: [
-									'3',
-									'1',
-									'0.0',
-									'133.0',
-									'0',
-									'5',
-								],
-							},
-						],
-					},
-					{
-						dimensions: [
-							'20190807',
-						],
-						metrics: [
-							{
-								values: [
-									'0',
-									'0',
-									'0.0',
-									'0.0',
-									'0',
-									'0',
-								],
-							},
-							{
-								values: [
-									'2',
-									'0',
-									'0.0',
-									'0.0',
-									'0',
-									'2',
-								],
-							},
-						],
-					},
-				],
-				totals: [
-					{
-						values: [
-							'128',
-							'34',
-							'52.94117647058824',
-							'141.6764705882353',
-							'0',
-							'238',
-						],
-					},
-					{
-						values: [
-							'100',
-							'23',
-							'65.21739130434783',
-							'63.34782608695652',
-							'0',
-							'155',
-						],
-					},
-				],
-				minimums: [
-					{
-						values: [
-							'0',
-							'0',
-							'0.0',
-							'0.0',
-							'0',
-							'0',
-						],
-					},
-					{
-						values: [
-							'0',
-							'0',
-							'0.0',
-							'0.0',
-							'0',
-							'0',
-						],
-					},
-				],
-				maximums: [
-					{
-						values: [
-							'11',
-							'5',
-							'100.0',
-							'444.0',
-							'0',
-							'36',
-						],
-					},
-					{
-						values: [
-							'9',
-							'3',
-							'100.0',
-							'402.0',
-							'0',
-							'20',
-						],
-					},
-				],
-			},
-		},
-	],
-};
